# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agentsfleet Control Plane API",
    "description": "API for managing workspaces, fleets, triggers, and runs.",
    "contact": {
      "name": "agentsfleet",
      "url": "https://agentsfleet.net"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.agentsfleet.net",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/v1/runners/me/memory/{fleet_id}": {
      "get": {
        "tags": [
          "Memory"
        ],
        "summary": "Load what a fleet remembers",
        "description": "The memory a fleet carries between runs, read at the start of one. The runner names the fleet, and the lease it holds is what makes that name legitimate. ",
        "operationId": "runner_hydrate_memory",
        "parameters": [
          {
            "name": "fleet_id",
            "in": "path",
            "description": "`UUIDv7` of the fleet.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryHydrateResponse"
                }
              }
            }
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      },
      "post": {
        "tags": [
          "Memory"
        ],
        "summary": "Capture what a fleet learned",
        "description": "Writes back what the run decided is worth keeping. The reply names what was stored and what was skipped, so a runner learns which of its entries did not survive the bounds. ",
        "operationId": "runner_capture_memory",
        "parameters": [
          {
            "name": "fleet_id",
            "in": "path",
            "description": "`UUIDv7` of the fleet.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryCaptureResponse"
                }
              }
            }
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MemoryCaptureResponse": {
        "type": "object",
        "description": "`POST /v1/runners/me/memory/{fleet_id}` reply — what the write did.",
        "required": [
          "stored",
          "skipped"
        ],
        "properties": {
          "skipped": {
            "type": "integer",
            "description": "Deltas refused for shape, which the runner should investigate.",
            "minimum": 0
          },
          "stored": {
            "type": "integer",
            "description": "Deltas written, after upsert.",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "MemoryDelta": {
        "type": "object",
        "description": "One durable memory item — the unit of both reading and writing.",
        "required": [
          "key",
          "content",
          "category"
        ],
        "properties": {
          "category": {
            "type": "string",
            "description": "Retention category, which decides eviction order."
          },
          "content": {
            "type": "string",
            "description": "The remembered content."
          },
          "key": {
            "type": "string",
            "description": "Stable key. A repeated key overwrites rather than accumulating."
          }
        },
        "additionalProperties": false
      },
...
      "MemoryHydrateResponse": {
        "type": "object",
        "description": "What a fleet remembers, compacted to fit one window.",
        "required": [
          "memory"
        ],
        "properties": {
          "memory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoryDelta"
            },
            "description": "The window's items."
          }
        },
        "additionalProperties": false
      },
      "MemoryPushRequest": {
        "type": "object",
        "description": "`POST /v1/runners/me/memory/{fleet_id}` request.",
        "required": [
          "lease_id",
          "fencing_token",
          "memory"
        ],
        "properties": {
          "fencing_token": {
            "type": "integer",
            "format": "int64",
            "description": "Monotonic guard; a reclaimed holder is rejected.",
            "minimum": 0
          },
          "lease_id": {
            "type": "string",
            "description": "The lease authorizing this write."
          },
          "memory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoryDelta"
            },
            "description": "The items to remember."
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Obtain a token via the CLI auth flow (POST /v1/auth/sessions) or GitHub OAuth"
      }
    }
  }
}